Add alreadyrunning switch for existing postgres and less log in mvn

+++ valinta-tulos-valintarekisteri-db/src/main/scala/fi/vm/sade/valintatulosservice/config/ITPostgres.scala
@@ -4,12 +4,18 @@ import java.io.File
 import java.nio.file.Files
 
 import fi.vm.sade.utils.slf4j.Logging
-import fi.vm.sade.utils.tcp.PortChooser
+import fi.vm.sade.utils.tcp.{PortChooser, PortFromSystemPropertyOrFindFree}
 import org.apache.commons.io.FileUtils
 
 import scala.sys.process.stringToProcess
 
 class ITPostgres(portChooser: PortChooser) extends Logging {
+
+  private def postgresAlreadyRunning(): Boolean = {
+    val alreadyRunning = System.getProperty("valintatulos.it.postgres.alreadyrunning")
+    (alreadyRunning != null && "true".equals(alreadyRunning))
+  }
+
   val port = portChooser.chosenPort
   val dataDirName = s"valintarekisteri-it-db/$port"
   val dbName = "valintarekisteri"
@@ -17,13 +23,18 @@ class ITPostgres(portChooser: PortChooser) extends Logging {
   val startStopRetryIntervalMillis = 100
   private val dataDirFile = new File(dataDirName)
   val dataDirPath = dataDirFile.getAbsolutePath
-  if (!dataDirFile.isDirectory) {
-    logger.info(s"PostgreSQL data directory $dataDirPath does not exist, initing new database there.")
-    Files.createDirectories(dataDirFile.toPath)
-    runBlocking(s"chmod 0700 $dataDirPath")
-    runBlocking(s"initdb -D $dataDirPath --no-locale")
+  if(!postgresAlreadyRunning) {
+    if (!dataDirFile.isDirectory) {
+      logger.info(s"PostgreSQL data directory $dataDirPath does not exist, initing new database there.")
+      Files.createDirectories(dataDirFile.toPath)
+      runBlocking(s"chmod 0700 $dataDirPath")
+      runBlocking(s"initdb -D $dataDirPath --no-locale")
+    }
+    logger.info(s"Using PostgreSQL in port $port with data directory $dataDirPath")
+  } else {
+    logger.info(s"Using already running PostgreSQL in port $port")
   }
-  logger.info(s"Using PostgreSQL in port $port with data directory $dataDirPath")
+
 
   private def isAcceptingConnections(): Boolean = {
     runBlocking(s"pg_isready -q -t 1 -h localhost -p $port -d $dbName", failOnError = false) == 0
@@ -53,43 +64,54 @@ class ITPostgres(portChooser: PortChooser) extends Logging {
   }
 
   def start() {
-    readPid match {
-      case Some(pid) => {
-        logger.debug(s"PostgreSQL pid $pid is found in pid file, not touching the database.")
-      }
-      case None => {
-        logger.info(s"PostgreSQL pid file cannot be read, starting:")
-        s"postgres --config_file=postgresql/postgresql.conf -D $dataDirPath -p $port".run()
-        if (!tryTimes(startStopRetries, startStopRetryIntervalMillis)(isAcceptingConnections)) {
-          throw new RuntimeException(s"postgres not accepting connections in port $port after $startStopRetries attempts with $startStopRetryIntervalMillis ms intervals")
+    if(!postgresAlreadyRunning) {
+      readPid match {
+        case Some(pid) => {
+          logger.info(s"PostgreSQL pid $pid is found in pid file, not touching the database.")
         }
-        runBlocking(s"dropdb -p $port --if-exists $dbName")
-        runBlocking(s"createdb -p $port $dbName")
-        runBlocking(s"psql -h localhost -p $port -d $dbName -f postgresql/init_it_postgresql.sql")
-
-        Runtime.getRuntime.addShutdownHook(new Thread(new Runnable {
-          override def run() {
-            stop()
+        case None => {
+          logger.info(s"PostgreSQL pid file cannot be read, starting:")
+          s"postgres --config_file=postgresql/postgresql.conf -D $dataDirPath -p $port".run()
+          if (!tryTimes(startStopRetries, startStopRetryIntervalMillis)(isAcceptingConnections)) {
+            throw new RuntimeException(s"postgres not accepting connections in port $port after $startStopRetries attempts with $startStopRetryIntervalMillis ms intervals")
           }
-        }))
+
+          initDb()
+
+          Runtime.getRuntime.addShutdownHook(new Thread(new Runnable {
+            override def run() {
+              stop()
+            }
+          }))
+        }
       }
+    } else {
+      initDb()
     }
   }
 
+  private def initDb() = {
+    runBlocking(s"dropdb -p $port --if-exists $dbName")
+    runBlocking(s"createdb -p $port $dbName")
+    runBlocking(s"psql -h localhost -p $port -d $dbName -f postgresql/init_it_postgresql.sql")
+  }
+
   def stop() {
-    readPid match {
-      case Some(pid) => {
-        logger.info(s"Killing PostgreSQL process $pid")
-        runBlocking(s"kill -s SIGINT $pid")
-        if (!tryTimes(startStopRetries, startStopRetryIntervalMillis)(() => readPid.isEmpty)) {
-          logger.error(s"postgres in pid $pid did not stop gracefully after $startStopRetries attempts with $startStopRetryIntervalMillis ms intervals")
+    if(!postgresAlreadyRunning) {
+      readPid match {
+        case Some(pid) => {
+          logger.info(s"Killing PostgreSQL process $pid")
+          runBlocking(s"kill -s SIGINT $pid")
+          if (!tryTimes(startStopRetries, startStopRetryIntervalMillis)(() => readPid.isEmpty)) {
+            logger.error(s"postgres in pid $pid did not stop gracefully after $startStopRetries attempts with $startStopRetryIntervalMillis ms intervals")
+          }
         }
+        case None => logger.info("No PostgreSQL pid found, not trying to stop it.")
       }
-      case None => logger.info("No PostgreSQL pid found, not trying to stop it.")
-    }
-    if (dataDirFile.exists()) {
-      logger.info(s"Nuking PostgreSQL data directory $dataDirPath")
-      FileUtils.forceDelete(dataDirFile)
+      if (dataDirFile.exists()) {
+        logger.info(s"Nuking PostgreSQL data directory $dataDirPath")
+        FileUtils.forceDelete(dataDirFile)
+      }
     }
   }
 }
